+++ src/DevilDaggersInfo.Tools/Extensions/AssetTypeExtensions.cs
@@ -1,0 +1,17 @@
+using DevilDaggersInfo.Core.Asset;
+using System.Numerics;
+
+namespace DevilDaggersInfo.Tools.Extensions;
+
+public static class AssetTypeExtensions
+{
+	public static Vector4 GetColor(this AssetType assetType) => assetType switch
+	{
+		AssetType.Audio => new(1, 0.25f, 1, 1),
+		AssetType.ObjectBinding => new(0.25f, 1, 1, 1),
+		AssetType.Mesh => new(1, 0.25f, 0.25f, 1),
+		AssetType.Shader => new(0.25f, 1, 0.25f, 1),
+		AssetType.Texture => new(1, 0.66f, 0.25f, 1),
+		_ => Vector4.One,
+	};
+}